+++ front/components/atoms/Logo/Logo.type.ts
@@ -1,4 +1,4 @@
-import { Mode } from '@common/types/Mode';
+import { Mode } from "@common/types/mode";
 
 export interface LogoProps {
   mode?: Mode;
