+++ front/components/molecules/NavMenu/NavMenu.type.ts
@@ -1,4 +1,4 @@
-import { Mode } from "@common/types/Mode";
+import { Mode } from "@common/types/mode";
 
 export interface NavMenuProps {
   mode?: Mode;
